requirements.txt
django
djangorestframework
django-extensions
gunicorn==21.2.0
whitenoise
djangorestframework-authtoken